+++ routes/web.php
@@ -30,28 +30,37 @@ use App\Models\User;
 Route::get('/', [ListingController::class, 'index']);
 
 // Show create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store listing Data
-Route::post("/listings", [ListingController::class, 'store']);
+Route::post("/listings", [ListingController::class, 'store'])->middleware('auth');
 
 // Show Edit Form
-Route::get("/listings/{listing}/edit", [ListingController::class, 'edit']);
+Route::get("/listings/{listing}/edit", [ListingController::class, 'edit'])->middleware('auth');
 
 // Edit Submit to update 
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Register user
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Create new user
 Route::post("/users", [UserController::class, 'store']);
 
 // Log user out
-Route::get("/logout", [UserController::class, 'logout']);
+Route::post("/logout", [UserController::class, 'logout'])->middleware('auth');
+
+// Show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+// Submit user route
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
+// Manage listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Get single listing from the controller
 Route::get("/listings/{id}", [ListingController::class, 'show']);
